[py] Move all test utility code into a single module.
No point in having multiple modules for so little code.
